Create Stephen chizimuzo favour 2024714065
This python program implements a grading system using 'if' and 'elif' statements. It takes a students score as input and prints the corresponding grade based on the following scale:
- 70 - 100 : Grade A
- 60 - 69 : Grade B
- 50 - 59 : Grade C
- 45 - 49 : Grade D
- 40 - 44 : Grade E
- 0 - 39 : Grade F
For each valid score, the program also prints'YES' to confirm the grading was successful. If the score is outside the valid range, it prints 'invalid score' and "No".
Created by: Stephen chizimuzo favour 
Reg No : 2024714065